backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# File processing
pypdf2==3.0.1
python-docx==1.1.0
extract-msg==0.41.0
email-parser==1.0.0
pandas==2.1.3
Pillow==10.1.0
pytesseract==0.3.10

# Text processing
python-magic==0.4.27
chardet==5.2.0

# Similarity and hashing
python-Levenshtein==0.23.0
difflib

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Async
aiofiles==23.2.1

# Vector Database
qdrant-client==1.7.0

# Embeddings
openai==1.12.0
numpy==1.26.2

